Relocating 'Lobby.java'
